backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
beautifulsoup4==4.12.3
lxml==5.2.1
python-pptx==0.6.23
